requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
openenv-core>=0.2.0
